+++ CHANGES.md
@@ -4,13 +4,18 @@
 
 - ...
 
+# v0.7.0
+
+- Add sendMsg and toChild function to component helpers for better messaging
+- Fix a test and coverage in worker helpers
+
 # v0.6.8
 
 - Add fetch value option for generic inputs
 
 # v0.6.7
 
-- Fix edge case with child components with synamic parents
+- Fix edge case with child components with dynamic parents
 
 # v0.6.6